requirements.txt
flask
streamlit
openai-whisper
requests
google-cloud-texttospeech